Convert to FP32

+++ algorithms/cssComm/cssCommAlgorithm.h
@@ -11,22 +11,22 @@ inline constexpr std::size_t kMaxNumChebyPolys = 32;
  CSS interface*/
 class CssCommAlgorithm final {
    public:
-    std::array<double, MAX_NUM_CSS_SENSORS> update(std::array<double, MAX_NUM_CSS_SENSORS> inputValues);
+    std::array<float, MAX_NUM_CSS_SENSORS> update(std::array<float, MAX_NUM_CSS_SENSORS> inputValues);
 
     void setNumSensors(uint32_t numberOfSensors);
     uint32_t getNumSensors() const;
-    void setMaxSensorValue(double maxValue);
-    double getMaxSensorValue() const;
+    void setMaxSensorValue(float maxValue);
+    float getMaxSensorValue() const;
     void setChebyCount(uint32_t count);
     uint32_t getChebyCount() const;
-    void setChebyPolynomials(const std::array<double, kMaxNumChebyPolys>& polynomials);
-    std::array<double, kMaxNumChebyPolys> getChebyPolynomials() const;
+    void setChebyPolynomials(const std::array<float, kMaxNumChebyPolys>& polynomials);
+    std::array<float, kMaxNumChebyPolys> getChebyPolynomials() const;
 
    private:
     uint32_t numSensors{};                                  //!< The number of sensors we are processing
-    double maxSensorValue{};                                //!< Scale factor to go from sensor values to cosine
+    float maxSensorValue{};                                //!< Scale factor to go from sensor values to cosine
     uint32_t chebyCount{};                                  //!< Count on the number of chebyshev polynomials we have
-    std::array<double, kMaxNumChebyPolys> chebyPolynomials{};   //!< Chebyshev polynomials to fit output to cosine
+    std::array<float, kMaxNumChebyPolys> chebyPolynomials{};   //!< Chebyshev polynomials to fit output to cosine
 };
 
 #endif
